+++ src/main/java/pl/kapelinski/mateusz/taco/cloud/tacos/TacoCloudApplication.java
@@ -1,4 +1,4 @@
-package pl.kapelinski.mateusz.taco.cloud.tacocloud;
+package pl.kapelinski.mateusz.taco.cloud.tacos;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
